+++ packages/picasso-forms/src/Form/story/index.jsx
@@ -218,3 +218,7 @@ however, you may need custom validators for more complex types of fields.
     title: 'No scrolling case',
     description: "Showcase Form's behavior on form submission error."
   }) // picasso-skip-visuals
+  .addExample('Form/story/FieldRequirements.example.tsx', {
+    title: 'Field requirements',
+    description: 'Showcase how to display field requirements.'
+  }) // picasso-skip-visuals