+++ packages/picasso/src/index.ts
@@ -190,6 +190,11 @@ export { default as RichText } from './RichText'
 export type { RichTextProps, ASTType } from './RichText'
 export { default as PasswordInput } from './PasswordInput'
 export type { PasswordInputProps } from './PasswordInput'
+export { default as FieldRequirements } from './FieldRequirements'
+export type {
+  FieldRequirementsProps,
+  FieldRequirement
+} from './FieldRequirements'
 
 // hygen code generator inserts export statements above this comment.
 export * from './Icon'