+++ packages/picasso/CHANGELOG.md
@@ -1,5 +1,29 @@
 # Change Log
 
+## 18.13.0
+
+### Minor Changes
+
+- [#2485](https://github.com/toptal/picasso/pull/2485) [`b8cf916d`](https://github.com/toptal/picasso/commit/b8cf916db7b4e64feb4afda3f1488690f0aefee4) Thanks [@OndrejTuma](https://github.com/OndrejTuma)! - ---
+
+  ### Link
+
+  - Aligned color for visited state with BASE. Changed to `#6727cf`
+
+## 18.12.0
+
+### Minor Changes
+
+- [#2476](https://github.com/toptal/picasso/pull/2476) [`c2e71484`](https://github.com/toptal/picasso/commit/c2e714841c17d90cf075630a8c4bffab92741ff5) Thanks [@ozgurkececioglu](https://github.com/ozgurkececioglu)! - ---
+
+  ### FieldRequirements
+
+  - add new input component to list requirements
+
+  ### FormField
+
+  - add new prop to pass field requirements as a ReactNode
+
 ## 18.11.0
 
 ### Minor Changes