+++ packages/topkit-analytics-charts/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Change Log
 
+## 28.0.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @toptal/picasso-charts@31.0.0
+
+## 27.0.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @toptal/picasso-charts@30.0.0
+
 ## 26.0.0
 
 ### Patch Changes